added a secondary allocator for small allocations, ran it through valgrind and fixed all memory bugs, features and bugfixes, and more tests

+++ internal/virtalloc/checksum.h
@@ -3,6 +3,6 @@
 
 #include <stddef.h>
 
-int get_checksum(size_t length, const char data[static length]);
+size_t get_checksum(size_t length, const char data[static length]);
 
 #endif
